embedding/pyproject.toml
[tool.poetry]
name = "embedding"
version = "0.1.0"
description = ""
authors = ["Joshua Bassin <josh.r.bassin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
# fastembed = "^0.3.4"
qdrant-client = {extras = ["fastembed"], version = "^1.10.1"}
pyyaml = "^6.0.2"
fastapi = "^0.112.0"
uvicorn = "^0.30.6"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
